requirements.txt
fastapi[all]
uvicorn
gunicorn
arxiv
elasticsearch
sentence-transformers
spacy
pytextrank
pymysql
aiomysql
sqlalchemy
aiofiles
pdfminer.six
python-dotenv
pydantic
jinja2
passlib[bcrypt]
apscheduler
accelerate
transformers
cachetools
uvloop
hf_xet